Fix up eventing to avoid leaks
Some zoomend listeners attached to global map
were not being properly cleaned up when the vector layer
is removed.  If the layer is being added and removed
multiple times (or even if the user zooms repeatedly),
the layer features will leak and at high scale,
the browser OOMs.  Added some more eventing propagation
on layer removal to handle this case.